requirements.txt
PyYAML==5.4.1
pytest==6.1.2
allure-pytest==2.8.19
allure-python-commons==2.8.19
selenium==3.141.0
pytest-rerunfailures==9.1.1
pytest-ordering==0.6
python-dateutil==2.8.1
redis==3.5.3
pytest-xdist==2.1.0
paramiko==2.7.2
scp==0.13.3
fire==0.3.1
openpyxl==3.0.5
xmltodict==0.12.0
faker==4.14.1
dataclasses_json==0.5.2
uplink==0.9.2
Appium-Python-Client==1.0.2
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
loguru==0.5.3
logzero==1.7.0
jsonpath==0.82
jsonschema==3.2.0
mysql-connector==2.2.9
PyMySQL==0.10.1
reportlab==3.5.56
requests==2.25.1
retry==0.9.2
